+++ basana/core/helpers.py
@@ -15,11 +15,12 @@
 # limitations under the License.
 
 from decimal import Decimal
-from typing import Any, Coroutine, List, Optional, Set, Union
+from typing import Any, Awaitable, Callable, Dict, List, Optional, Union
 import asyncio
 import contextlib
 import decimal
 import logging
+import uuid
 import warnings
 
 import aiohttp
@@ -68,64 +69,109 @@ class TaskPool:
     """
     A class for managing a pool of asyncio tasks.
 
-    :param size: The maximum size of the task pool.
+    :param size: The maximum number of tasks to be running at the same time.
+    :param max_queue_size: The maximum number of coroutine functions to be waiting in the queue for execution.
     """
-    def __init__(self, size: int):
-        assert size > 0, "Invalid size"
-        self._max_size = size
-        self._tasks: Set[asyncio.Task] = set()
-        self._done: List[asyncio.Task] = []
+    def __init__(self, max_tasks: int, max_queue_size: Optional[int] = None):
+        assert max_tasks > 0, "Invalid max_tasks"
+        assert max_queue_size is None or max_queue_size > 0, "Invalid max_queue_size"
+
+        self._max_tasks = max_tasks
+        self._queue = LazyProxy(
+            lambda: asyncio.Queue(maxsize=max_tasks if max_queue_size is None else max_queue_size)
+        )
+        self._tasks: Dict[str, asyncio.Task] = {}
+        self._queue_timeout = 1.0
+        self._active = 0
 
     @property
-    def idle(self) -> int:
+    def idle(self) -> bool:
         """
-        True if there are no active tasks in the pool, False otherwise.
+        True if there are no coroutines being executed and there are no coroutines waiting in the queue,
+        False otherwise.
         """
-        return len(self._tasks) == 0
+        return self._active == 0 and self._queue.empty()
 
-    async def push(self, coroutine: Coroutine[Any, Any, Any]):
+    async def push(self, coroutine_func: Callable[[], Awaitable[Any]]):
         """
-        Adds a coroutine to the task pool. If the pool is full it will block until there is room for the new task.
+        Adds a coroutine function to the queue. It may block if the queue is full.
 
-        :param coroutine: The coroutine to be added to the task pool.
+        :param coroutine_func: The coroutine function to be added to the task pool.
         """
-        # Wait for some task to complete if there is no more room.
-        while len(self._tasks) >= self._max_size:
-            await self._wait_impl(timeout=None, return_when=asyncio.FIRST_COMPLETED)
-        self._tasks.add(asyncio.create_task(coroutine))
 
-    def pop_done(self) -> List[asyncio.Task]:
-        """
-        Returns the tasks that are done running since the last call to pop_done.
-        """
-        ret = self._done
-        self._done = []
-        return ret
+        await self._queue.put(coroutine_func)
+
+        # Create a new task if necessary.
+        idle_tasks = len(self._tasks) - self._active
+        if idle_tasks == 0 and len(self._tasks) < self._max_tasks:
+            task_name = uuid.uuid4().hex
+            task = asyncio.create_task(self._task_main(task_name))
+            # We check before registering the task because if eager tasks are enabled (Python >= 3.12) the task may
+            # have already ran by the time we got here.
+            if not task.done() and task_name not in self._tasks:
+                self._tasks[task_name] = task
 
     def cancel(self):
         """
-        Requests all tasks in the pool to be canceled.
+        Requests all tasks in the pool to be canceled and clears the queue.
         """
-        pending = [task for task in self._tasks if not task.done()]
-        for task in pending:
+        for task in self._tasks.values():
             task.cancel()
+
+        # Empty the queue.
+        while self._queue.qsize():
+            self._queue.get_nowait()
+            self._queue.task_done()
 
     async def wait(self, timeout: Optional[Union[int, float]] = None) -> bool:
         """
-        Waits for all tasks in the pool to complete. Returns True if all tasks are done, False otherwise.
+        Waits for all tasks in the pool to complete.
 
         :param timeout: The maximum number of seconds to wait for tasks to complete. If None, wait indefinitely.
+        :returns: Returns True if all the coroutines in the queue have been processed, False otherwise.
         """
-        return await self._wait_impl(timeout=timeout, return_when=asyncio.ALL_COMPLETED)
 
-    async def _wait_impl(self, timeout: Optional[Union[int, float]], return_when: str) -> bool:
-        done: List[asyncio.Task] = []
-        if self._tasks:
-            done, _ = await asyncio.wait(self._tasks, timeout=timeout, return_when=return_when)
-        for task in done:
-            self._tasks.remove(task)
-            self._done.append(task)
-        return len(done) > 0
+        ret = False
+        try:
+            await asyncio.wait_for(self._queue.join(), timeout=timeout)
+            ret = True
+        except asyncio.TimeoutError:
+            pass
+        return ret
+
+    async def _task_main(self, task_name: str):
+        current_task = asyncio.current_task()
+        # Register ourselves in the task registry if not already there. This happens with eager tasks (Python >= 3.12).
+        if current_task not in self._tasks:
+            assert current_task is not None
+            self._tasks[task_name] = current_task
+
+        try:
+            eof = False
+            while not eof:
+
+                try:
+                    coro_func = await asyncio.wait_for(self._queue.get(), timeout=self._queue_timeout)
+                except asyncio.TimeoutError:
+                    # This double check was introduced to workaround a race condition that was causing the above pop
+                    # to timeout even though there were items in the queue:
+                    # 2025-10-16 15:08:30 - DEBUG - basana.core.helpers - Task pop timed out {"total_tasks": 1, "active_tasks": 0, "queue_size": 2}  # noqa: E501
+                    # 2025-10-16 15:08:30 - DEBUG - basana.core.helpers - Task is about to exit {"total_tasks": 0, "active_tasks": 0, "queue_size": 2}  # noqa: E501
+                    eof = self._queue.empty()
+                    continue
+
+                try:
+                    self._active += 1
+                    await coro_func()
+                except Exception:
+                    # Individual coroutine failures should not take down a worker
+                    pass
+                finally:
+                    self._active -= 1
+                    self._queue.task_done()
+        finally:
+            # Remove ourselves from the task registry once we're done.
+            self._tasks.pop(task_name)
 
 
 @contextlib.contextmanager